requirements.txt
fastapi==0.118.0
google-auth==2.41.1
google-auth-oauthlib==1.2.2
gspread==6.2.1
openai==2.1.0
uvicorn==0.37.0
